+++ src/main/kotlin/com/github/denofevil/aurelia/Aurelia.kt
@@ -22,8 +22,9 @@ import com.intellij.psi.util.CachedValuesManager
 object Aurelia {
     val ICON = IconLoader.getIcon("/icons/aurelia-icon.svg", Aurelia::class.java)
 
-    private val PROPERTY_BINDINGS = listOf("bind", "one-way", "two-way", "one-time", "from-view", "to-view", "dispatch")
-    val INJECTABLE = listOf("delegate", "trigger", "call", "for", "ref") + PROPERTY_BINDINGS
+    val PROPERTY_BINDINGS = listOf("bind", "one-way", "two-way", "one-time", "from-view", "to-view")
+    val EVENT_BINDINGS = listOf("trigger", "delegate", "call")
+    val INJECTABLE = listOf("for", "ref") + PROPERTY_BINDINGS + EVENT_BINDINGS
     val REPEAT_FOR = listOf("repeat.for", "virtual-repeat.for", "reorderable-repeat.for")
     const val REPEAT_FOR_OF_KEYWORD = " of "