+++ CHANGELOG.md
@@ -6,6 +6,15 @@
 
 ### Added
 
+- Proper `repeat.for` detection and reference resolving
+- Ignoring binding behaviors and value converts for code injection
+- Suppressing of missing promise for aurelia js injected code
+- $this, $parent, $index, $event support for code injections
+
+## [2.1.0]
+
+### Added
+
 - Support for 251
 
 ## [2.0.1]